Create file in typescript
