+++ src/CreatePrivateGame/CreatePrivateGame.css
@@ -13,7 +13,7 @@
 
 .create-private-game {
   padding: 20px;
-  background-color: white;
+  background-color: rgb(212, 33, 182);
   border-radius: 5px;
   box-shadow: 0 2px 4px rgba(0, 0, 0, 0.2);
   color: black;
